Refactor extensibility logic Reuse handlers throughout class and instance proxies Move instances' proxying from constructors to function
