+++ src/config/Middleware/generateApplyHandler.js
@@ -1,0 +1,22 @@
+/* eslint-disable */
+const sortPlugins = require('./sortPlugins');
+const generateMiddlewaredFunction = require('./generateMiddlewaredFunction');
+
+module.exports = (namespace) => {
+    return function (origFunction, thisArg, originalArgs) {
+        const memberPluginsApply = window.plugins?.[namespace]?.['function']?.['apply'];
+        if (memberPluginsApply && !Array.isArray(memberPluginsApply)) {
+            throw new Error(`Expected Array in function/apply config section for ${namespace}`);
+        }
+
+        if (!memberPluginsApply) {
+            return origFunction.apply(thisArg, originalArgs);
+        }
+
+        return generateMiddlewaredFunction(
+            origFunction,
+            sortPlugins(memberPluginsApply),
+            thisArg
+        )(...originalArgs);
+    };
+}
