+++ src/config/Middleware/index.js
@@ -1,52 +1,7 @@
 /* eslint-disable */
-
-/**
- * Sort plugins by position:
- * @param {Array} plugins
- * @param {string} errorText
- */
-const sortPlugins = (plugins, errorText) => plugins
-    .sort((a, b) => {
-        if (a.position > b.position) return -1;
-        if (a.position < b.position) return 1;
-        throw new Error(errorText);
-    });
-
-/**
- * Get error text to show if two plugins are defined with same positions
- * @param {string} memberName
- * @param {string|null} className
- * @param {string} namespace
- */
-const getHandlerErrorText = (memberName, className, namespace) => {
-    return `Cannot have equal position on different plugins of the same ${memberName} member
-    on ${ className ? `${className} class` : '' } in ${namespace} namespace!`;
-};
-
-/**
- * Middlewaring given original member
- * @param {Function} origMember
- * @param {Array} sortedPlugins
- * @param Context origContext
- */
-function middlewareFunction(origMember, sortedPlugins, origContext) {
-    return function (...args) {
-        const starter = typeof origMember === 'function'
-            ? (...originalArgs) => origMember.apply(origContext, originalArgs)
-            : origMember;
-
-        const newMember = sortedPlugins.reduce(
-            (acc, { implementation }) => () => {
-                return typeof origMember === 'object'
-                    ? implementation(acc, origContext)
-                    : implementation(args, acc, origContext);
-            },
-            starter
-        );
-
-        return newMember(args);
-    };
-}
+const generateGetHandler = require('./generateGetHandler');
+const generateApplyHandler = require('./generateApplyHandler');
+const generateConstructHandler = require('./generateConstructHandler');
 
 /**
  * Middleware function is supposed to wrap source classes
@@ -56,80 +11,6 @@ function middlewareFunction(origMember, sortedPlugins, origContext) {
  */
 function middleware(Class, namespace) {
     Class.prototype.__namespace__ = namespace;
-
-    const applyHandler = function (origFunction, thisArg, originalArgs) {
-        const memberPluginsApply = window.plugins?.[namespace]?.['function']?.['apply'];
-        if (memberPluginsApply && !Array.isArray(memberPluginsApply)) {
-            throw new Error(`Expected Array in function/apply config section for ${namespace}`);
-        }
-
-        if (!memberPluginsApply) {
-            return origFunction.apply(thisArg, originalArgs);
-        }
-
-        return middlewareFunction(
-            origFunction,
-            sortPlugins(
-                memberPluginsApply,
-                getHandlerErrorText(origFunction.prototype.constructor.name, null, namespace)
-            ),
-            thisArg
-        )(...originalArgs);
-    };
-
-    const getHandler = function (target, memberName, rec) {
-        if (memberName === 'Symbol.iterator') {
-            return target[Symbol.iterator].bind(target);
-        }
-        const origMember = Reflect.get(target, memberName, rec);
-
-        const memberPluginsGet = window.plugins?.[namespace]?.['class']?.['get']?.[memberName];
-        if (!memberPluginsGet) {
-            return origMember;
-        }
-
-        const middlewaredFunction = middlewareFunction(
-            origMember,
-            sortPlugins(
-                memberPluginsGet,
-                getHandlerErrorText(memberName, Object.getPrototypeOf(target).constructor.name, namespace)
-            ),
-            target
-        );
-
-        if (typeof origMember === 'object') {
-            return middlewaredFunction();
-        }
-
-        return middlewaredFunction;
-    };
-
-    const constructHandler = function (TargetClass, args) {
-        const instance = new TargetClass(...args);
-        const namespacePluginsConstruct = window.plugins?.[namespace]?.['class']?.['construct'] || {};
-
-        Object.entries(namespacePluginsConstruct).forEach(
-            ([memberName, memberPluginsConstruct]) => {
-                const origMember = instance[memberName];
-                const sortedPlugins = sortPlugins(
-                    memberPluginsConstruct,
-                    getHandlerErrorText(memberName, TargetClass.name, namespace)
-                );
-
-                const newMember = sortedPlugins.reduce(
-                    (acc, { implementation }) => {
-                        return implementation(acc, instance);
-                    },
-                    origMember
-                );
-
-                instance[memberName] = newMember;
-            }
-        );
-
-        return instance;
-    };
-
     const handler = {};
 
     // All classes inherit from extensible classes
@@ -140,20 +21,20 @@ function middleware(Class, namespace) {
         Object.defineProperty(
             handler,
             'apply',
-            { value: applyHandler }
+            { value: generateApplyHandler(namespace) }
         );
     } else {
         // Get handler for members - intercepts `get` calls, meant for class static members
         Object.defineProperty(
             handler,
             'get',
-            { value: getHandler }
+            { value: generateGetHandler('class', namespace) }
         );
         // Construct handler for classes - intercepts `new` operator calls, changes properties
         Object.defineProperty(
             handler,
             'construct',
-            { value: constructHandler }
+            { value: generateConstructHandler(namespace) }
         );
     }
